Add various correctness fixes (see descriptions)
Fragmentation source was corrected by introducing factor of particle
velocity. Previously the units of the fragmentation source were
incorrect.

"u" mode iteration was improved by sampling various values of vent
pressure as a function of chamber pressure, and selecting the highest
pair of chamber pressures that bracket the sonic condition.

Reduced atol for iteration in "p" mode due to sensitivity of inlet
velocity to chamber pressure. Choice of atol at 1e-5 is a compromise
between closeness to the choking pressure (small atol) and having enough
robustness to the sonic termination condition (for which small atol may
lead to insufficiently long output solutions). The direction of initial
guessing with respect to atol was changed.

Flag to ignore deformation energy term in liquid phase was added
(default:False).

Extra overflow/illegal value handling was added to the viscosity source
function.