Change solution evaluation to solve_ivp's dense_output with const extrapolation and fix j0=0 case.
Due to numerical perturbation arising from evaluation on a mesh
different from the one used to solve the steady state IVP, the solution
may sometimes not be evaluable at the top boundary (choking occurs below
the top). The attempted solution is to use dense_output=True every time
solve_ivp from scipy is called. Due to a corner case bug in the scipy
routine, this is not possible. Instead, we use dense_output=True only
when extracting the final solution. This has the same effect as the
previous strategy: the solution may not be evaluable at the top
boundary. As a workaround, a constant extrapolation is used to replace
any out-of-bounds values near the top boundary with the last legitimate
value within the interval. This prevents extrapolating using the cubic
function in the Radau method implementation in solve_ivp, which
sometimes gives large negative extrapolated values.

The j0=0 case was fixed, replacing the erroneous yL value with yM.